Add sapling updates
Saplings now grow into trees identical to those created in world generation. They only grow if they have enough air space to grow, and are open to the sky.

They have a 50% chance of growing every 16 ticks. This probability will be decreased in the future.

Currently, they do not break when placed on stone, sand, or other such blocks. This will be changed in a future update.